requirements.txt
tqdm
pandas
datasets>=2.19.1
google-generativeai>=0.5.4
streamlit>=1.31.1
watchdog